% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acd-filter.R
\name{acdfilter}
\alias{acdfilter}
\title{Filtering new data}
\usage{
acdfilter = function(spec, data, out.sample = 0,  n.old = NULL, skew0 = NULL, shape0 = NULL, ...)
}
\arguments{
\item{spec}{An ACDspec object with fixed parameters}

\item{data}{dataset}

\item{n.old}{Number of old data that will be used to fit the model specification. The last value will be
used to start the filtering}
}
\value{
An ACDfiltering object
}
\description{
Filtering new data with regards to the model specification with fixed parameters
}
