% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acd-fitmethod.R
\name{acdfit}
\alias{acdfit}
\title{Model Estimation}
\usage{
acdfit = function(spec, data, solver = "ucminf", out.sample = 0, solver.control = list(),
      fit.control = list(stationarity = 0, fixed.se = 0, scale = 0,n.sim = 2000),
      skew0 = NULL, shape0 = NULL, cluster = NULL, ...)
}
\arguments{
\item{spec:}{ACDspec object with model specification}

\item{data:}{Data}

\item{solver:}{The chosen solver, which could be "ucminf", "optim", ...and "ms-" version that allows for multi-starts optimization strategies.
refers to the documentation of Alexios's racd for more detials}

\item{out.sample:}{The length of data that will be used as out of sample model evaluation}

\item{solver.control:}{List of control values for solvers. For examples, we could specify number of restarts that will be used in multi-strarts
optimization strategies}

\item{fit.control:}{List of fitting controls arguments, such as number of simulation to generate starting values of optimization routines}

\item{cluster:}{An makePSHOCKcluster object to using parallel in model estimation}
}
\value{
An ACDfit object which contains information about model estimation
}
\description{
Estimation ACD model with fitting and solver control arguments
}
