% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acd-path.R
\name{acdpath}
\alias{acdpath}
\title{ACD simulated path}
\usage{
acdpath = function(spec, n.sim = 1000, n.start = 0, m.sim = 1, presigma = NA,
       prereturns = NA, preresiduals = NA, preskew = NA, preshape = NA, rseed = NA,
       cluster = NULL, ...)
}
\arguments{
\item{spec}{A spec object with fixed parameters}

\item{pre..}{Must be supplied. The values of returns, residuals, skew, shape that will be used to start the simulation process}
}
\value{
A ACDpath object that will be a simulated process for an ACD process.
}
