% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acd-sim.R
\name{acdsim}
\alias{acdsim}
\title{ACD simulation}
\usage{
acdsim = function(fit, n.sim = 1000, n.start = 0, endpoint = NA, m.sim = 1,
           presigma = NA, prereturns = NA, preresiduals = NA, preskew = NA,
          preshape = NA, rseed = NA,...)
}
\arguments{
\item{fit}{ACDfit object}

\item{n.sim}{Simulation horizon}

\item{n.start}{Burn-in period in the simulation}

\item{endpoint}{This is my additional argument, specify the endpoint in the fitted series that will be used as
pre-values to start the simulation}

\item{m.sim}{Number of simulation serires}

\item{rseed}{The seed number that must be suplied at the same legth as m.sim arguments to ensure the reproducibility}

\item{pre..}{The user-specified pre-values of sigma, returns, residuals, skew, shape that will be used to start the simulation}
}
\value{
An ACDsim object with simulated values for returns, sigma, skew, shape
}
\description{
Produce simulation for ACD model.
}
