% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acd-specmethod.R
\name{acdspec}
\alias{acdspec}
\title{Model specification}
\usage{
function(variance.model = list(model = "gjrGARCH", garchOrder = c(1, 1),variance.targeting = FALSE),
      mean.model = list(armaOrder = c(0,0), include.mean = TRUE),
      distribution.model = list(model = "sged", skewOrder = c(1,0, 1),skewshock = 1, skewshocktype = 1, skewmodel = "quad",
                                shape1Order = c(1, 0, 1), shape1shock = 1, shape1shocktype = 1,shape1model = "quad",
                                 shape2Order = c(1, 0, 1), shape2shock = 1, shape2shocktype = 1,shape2model = "quad",exp.rate = 1),
      start.pars = list(), fixed.pars = list())
}
\arguments{
\item{variance.model}{Specification of variance process. Only support
sGARCH and gjrGARCH. Do not support external regessor.}

\item{mean.model}{Specification of mean process. Only support maximum ARMA(1,1)}

\item{distribution.model}{Specification of skew/shape distribution's
   parameter dynamids. Only support "sgt", "sged" and "nig"
as conditional distribution.}

\item{start.pars}{List of starting values for specific parameters using in optimization routines}

\item{fixed.pars}{List of parameters that we want to fixed, i.e, drop out of optimization routines}
}
\value{
Return an ACDspec,i.e., specification objects.
}
\description{
Specify models to estimate with mean, variance and higher moment equations
}
