% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acd-disbounds.R
\name{exptransform1}
\alias{exptransform1}
\title{Exponential Transformation}
\usage{
exptransform(x,lower,rate,inverse = FALSE)
}
\arguments{
\item{x}{value to be transfromed}

\item{lower:}{Lower bound}

\item{rate:}{exponential rate. The default is 1}

\item{inverse:}{Logical value. The default is FALSE}
}
\description{
Exponential-transform or Inverse Exponential transform w.r.t specify LB and UB bounds
}
