% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acd-disbounds.R
\name{logtransform}
\alias{logtransform}
\title{Log Transformation}
\usage{
logtransform(x,lower,upper,inverse = FALSE)
}
\arguments{
\item{x}{value to be transfromed}

\item{lower:}{Lower bound}

\item{upper:}{Uper bound}

\item{inverse:}{Logical value. The defaul is FALSE}
}
\description{
logtransform or Inverse-logtransform w.r.t specify LB and UB bounds
}
