% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acd-helpmethods.R
\name{setboundsacd}
\alias{setboundsacd}
\title{Set bounds for conditional density distribution}
\usage{
setboundsacd(object, value)
}
\arguments{
\item{object}{The model specification object}
}
\description{
Set the bounds to the shape parameters of conditional distribution to which the density is exist
}
