% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acd-helpmethods.R
\name{setfixedacd}
\alias{setfixedacd}
\title{Set fixed parameters}
\usage{
setfixedacd(object, value)
}
\arguments{
\item{object}{Model sepecification}

\item{value}{A named list of parameters to be fixed}
}
\description{
Set fixed parameters to a model specification
}
