% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acd-helpmethods.R
\name{setstartacd}
\alias{setstartacd}
\title{Set starting values for a model specification}
\usage{
setstartacd(object, value)
}
\arguments{
\item{object}{Model specification class}

\item{Value}{A named list of starting values for parameters, which are then put to the estimation procedures}
}
\description{
Set the starting values for estimation routine of a model specification objects
}
