% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acd-disbounds.R
\docType{data}
\name{TinY}
\alias{TinY}
\title{Distribution bounds}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
distbounds(distribution)
}
\arguments{
\item{distribution:}{a string input with three supported conditional distribution: sgt, sged and sst}
}
\value{
a vector with the Lowerbounds(LB), Upperbounds(UB) for skew and shape parameters to which the denstiy exist
}
\description{
Returns the bounds to which a distribution density exist
}
\keyword{datasets}
